[refactor] wikidata: cache wikidata properties in searxng data
