+++ .github/workflows/data-update.yml
@@ -31,7 +31,7 @@ jobs:
           - update_external_bangs.py
           - update_firefox_version.py
           - update_engine_traits.py
-          - update_wikidata_units.py
+          - update_wikidata.py
           - update_engine_descriptions.py
 
     permissions:
